Implemented Password Setup on LCD
Allowed for the password setup system to be displayed on the LCD. Now you can go through the program from setting up the time, date, and password to displaying the home screen.  Once at home press enter (#) to enter the password you saved and see the menu. Press the home key (*) to return home.