minor change
allows each key pressed in sensor menu to reset idle count.

TODO: Modify PWM structure, so that ADC update a global duty cycle for pwm that will be  called in the begnning.

Verify hall sensor, sensitivity?

modify sensor menu so that it reports live status of the sensors

NMOSFET for LEDs